Closes #17: Use evaluateJavascript instead of  urls.
The issue was that we called `loadUrl`, but then we immediately called
`loadUrl` with a `javascript:` url, which apparently cancelled the load
on later API levels.

`evaluateJavascript` was introduced in API 19, less than our min SDK,  but
isn't available in AmazonWebView so we were unable to use it. However,
we removed AWV in #2 so we're free to use it.