Closes #51: Set toolbar navigation button state for page shown.
The back buttons break when you visit the home page so I can't
thoroughly test this: #57 should fix the back behavior.

It was immediately clear to me so I just want to point out that the initial
toolbar state will be correct because `setup` has the same effect as
calling `invalidateActions` immediately.