Issue #54: Convert AndroidManifest to non-leanback.
We removed the TV-centric `<uses-features` tags, the app banner (used in
leanback launch screens), and switched the launcher to a non-leanback
launcher (at least one launcher is required).

Android Studio is warning me that the manifest expects TV-centric tags
(e.g. leanback_launcher) but I think this may be residual data in my
project (I tried removing .idea but it didn't help) or the fact that we
include the leanback support library (for our homescreen grid). It
hasn't stopped me from building so I think it will be okay.