Closes #214: Remove DOMFocusedElementCache.
This hack was added to Fire TV because, when the WebView loses and
regains focus, it forgets which DOM element was focused and, because
we're using dpad navigation, it's important the correct element is
focused. It's no longer necessary here and removing it fixes this bug:
it's unclear why this was causing problems because it doesn't actually
mutate the DOM's state in any way.