Issue #58: Simplify show/hide homeScreen; remove telemetry.
On Fire TV, the `showHomeScreenByUser` (or whatever) method was intended
to abstract users away from the fact that they may need to add
telemetry. However, it was often called incorrectly so I'm changing the
structure that we'll track the clicks, rather than hide/show so that we
don't run into similar problems. This has the added benefit that we can
capture these events through onNavigationEvent.