Closes #374: Correct and clean up startup time measurement.
The original implementation is
https://github.com/mozilla-mobile-skunkworks/firefox-connect/pull/352/

This corrects the original implementation by moving the initial
measurement from LocaleAwareApplication.onCreate to
FocusApplication.onCreate; otherwise it is functionally equivalent.

It cleans up the original implementation by encapsulating all of the
functionality in a single class.

This implementation is less safe - TelemetryWrapper.startupCompleteEvent
can be called directly to record multiple events - but I don't have time
to correct that.