Closes #454: Move BrowserTouchExplorationStateChangeListener to onCreateView lifecycle.
A crash was caused because the homeScreen view was null in when the listener
was called. The view was null because, presumably, onCreateView hadn't been
called yet. To fix this issue, I attach the listener in
onCreateView to ensure the view is not null.

The listener currently only updates UI state so there is no need to
attach it before onCreateView where the UI would not exist anymore.

I like that this change is more correct but I don't like how much code
it adds.