Issue #540: Remove WebView from ClearDataFragment.
In #433, we learned we cannot have two WebView instances on these
devices or else we crash.

This is a speculative fix: I was unable to reproduce the primary STR in
this bug, however, I was able to reproduce a crash on a similar STR also
mentioned in this bug and I believe they have the same root cause.