Issue #75: Simplify fullscreen layoutParams.
We had to set an explicit height on the layout params when the WebView
was contained in a NestedScrollView (i.e. the height of it is larger
than the full screen so MATCH_PARENT will make the fullscreen container
larger than the screen) but now that it is no longer contained in a
scrolling container, we can simplify it to match_parent.