Issue #117: Move home tiles to new NavigationOverlayFragment.
A fragment implementation allows us to decouple and better encapsulate
the navigation overlay experience from the browser experience.

This refactor breaks some previous functionality including animations
and a11y: it's not worth restoring this functionality considering it
will need to be rewritten when we use the new layout in an upcoming
commit.