Issue #117: Add UnpinTilesTest.
I wanted to write a test to ensure we can't remove all of the home tiles
because I don't handle that case in the navigation overlay. While I was in the
area, I made the test handle unpinning in a variety of cases.

This is unlikely to be all of the test code added for this issue.