Issue #153: Remove TouchExplorationStateChangeListener in BrowserAppBarLayoutController.
I verified the memory leak is fixed by enabling don't keep activities
and checking if StrictMode logs a multiple instances exception when
pressing home and returning to the app from the app launcher.

Unfortunately, #167 prevents the BrowserAppBarLayoutController from
properly modifying the scroll state of the layout so we cannot manually
confirm the layout is working correctly after this change. However, I
rewrote the change on an earlier commit and confirmed the behavior is
working as expected.