Issue #167: Add BrowserAppViewModel and tests.
This will let us decouple the isToolbarScrollEnabled view state from the
BrowserAppBarLayoutController, clarifying the code (e.g. view state
is always available via the LiveData), and allowing us to test the view
logic. This implementation is not an ideal state: state should be pushed
from the model but we don't have a distinct model yet. :)