Closes #190: Don't assert nav overlay state before changing vis.
The assertion was causing the crash.

I noticed the current code is not explicit and confusing: `loadUrl`
unexpectedly hides the nav overlay. It tries to funnel the code down the
same path but, instead, I think each user action that wants to close the
nav overlay should have an explicit call to hide the overlay. However,
the ScreenController code and session initialization makes this complex
to change currently. As such, the best way to fix the crash was to
remove the assertions.

More generally, I think the assertions are actually not useful:
callers state the desired view state ("when a home tile is pressed, I
want the overlay to be closed") so it adds complexity to sometimes
need to check if the view is in some state before making the call, to
avoid triggering the assertions (e.g. if we're toolbar on startup,
we want to dismiss the nav overlay but don't want to on non-startup).