Issue #213: Add DefaultSearchOnEditorActionListener.
Once the search is submitted, this listener will be used to make the
actual search.

This code is adapted from `SearchOnGoEditorActionListener` and the
`setKeyboardListener` method that attaches it to the EditText. That
existing code is a little more generic whereas this code is a little
simpler because it's specific to the use case.