Issue #219: Create ServiceLocator with static factory function.
We don't want the constructor to have side effects (e.g. it makes
mocking/testing more difficult) so we put that initialization code in
the `init` method. However, it's easy for the caller to forget to call
this so we wrap it all in a static factory function and make the
constructor private.