Issue #186: Move restoration check to IntentResponder from IntentValidator.
We want to handle the intent redelivery outside of where we extract
values from the Intent. Also, IntentValidator is going to be converted to
more appropriate extension functions

This is a slight behavior change over the previous code: we only check for
intent redelivery in onCreate now. However, this is how the code was
before it was refactor in the last PR:
https://github.com/mozilla-mobile/firefox-echo-show/pull/249