Issue #305: Remove AppBar elevation assignment; split semiOpaqueBackground.
RCA: setting an elevation on the AppBar will pause playing WebView
videos when the AppBar is scrolled off-screen for an unknown reason. See
XML for details.

One side-effect of this change is that the AppBar displays a shadow on
the initial homescreen on all devices: I didn't feel it was worth the
complexity to selectively enable this fix on different devices to avoid
this on all devices but the ES5.